Remove `PROPOSAL_MODULES_COUNT` state.
Storing this led to a state inconsistency when we would update it's
value when the update_proposal_modules method was called. At that
point we would update its value, but the new modules would not
actually be installed as the submessages to add them had not been
executed.

It turns out that we can enforce that there is always a proposal
module in constant time without having that state. This just removes
it in favor of that method.